+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-### 0.2.13 (Upcoming Release)
+### 0.2.14 (Upcoming Release)
+
+### 0.2.13
 
 - Minimum Spanning Tree accepts distance matrix encapsulated by MstProblem as input.
 - Export lower and upper bounds of the game and strategy type (Pure or Mixed) in Game solver.